requirements.txt
pwntools==4.10.0
pyelftools==0.29